# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /users:
    get:
      tags:
        - users-controller
      summary: Отобразить всех пользователей
      operationId: findAllUsers
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - users-controller
      summary: Создать нового пользователя
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        201:
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /users/{id}:
    get:
      tags:
        - users-controller
      summary: Отобразить пользователя по ID
      operationId: findUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'

    put:
      tags:
        - users-controller
      summary: Изменить информацию о пользователе по ID
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Пользователь успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'

    delete:
      tags:
        - users-controller
      summary: Удалить пользователя по ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        204:
          description: Пользователь успешно удален
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
  /achievements/{personId}:
    get:
      tags:
        - achievements-controller
      summary: Найти достижение по ID человека
      operationId: findByPersonId
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        200:
          description: Список достижений человека
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AchievementResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
  /outstandingPeople:
    get:
      tags:
        - outstanding-people-controller
      summary: Отобразить всех выдающихся сотрудников
      operationId: findAllOutstandingPeople
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutstandingPeopleResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - outstanding-people-controller
      summary: Создать запись о новом сотруднике
      operationId: createOutstandingPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutstandingPersonCreateRequest'
      responses:
        201:
          description: Запись успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutstandingPersonResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /outstandingPeople/{id}:
    get:
      tags:
        - outstanding-people-controller
      summary: Отобразить сотрудника по его ID
      operationId: findOutstandingPersonById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        200:
          description: Найденный выдающийся человек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutstandingPersonResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'

    put:
      tags:
        - outstanding-people-controller
      summary: Изменить информацию о сотруднике по ID
      operationId: updateOutstandingPerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutstandingPersonUpdateRequest'
      responses:
        200:
          description: Запись успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutstandingPersonResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /outstandingPeople/search/byName:
    get:
      tags:
        - outstanding-people-controller
      summary: Отобразить сотрудников по имени
      operationId: findOutstandingPersonByName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Имя для поиска
          example: "Иван"
      responses:
        200:
          description: Список найденных выдающихся людей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutstandingPeopleResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /outstandingPeople/search/bySurname:
    get:
      tags:
        - outstanding-people-controller
      summary: Отобразить сотрудников по фамилии
      operationId: findOutstandingPersonBySurname
      parameters:
        - name: surname
          in: query
          required: true
          schema:
            type: string
          description: Фамилия для поиска
          example: "Петров"
      responses:
        200:
          description: Список найденных выдающихся людей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutstandingPeopleResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /outstandingPeople/search/byNameAndSurname:
    get:
      tags:
        - outstanding-people-controller
      summary: Отобразить сотрудника по имени и фамилии
      operationId: findOutstandingPersonByNameAndSurname
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Имя для поиска
          example: "Иван"
        - name: surname
          in: query
          required: true
          schema:
            type: string
          description: Фамилия для поиска
          example: "Петров"
      responses:
        200:
          description: Список найденных выдающихся людей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutstandingPeopleResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
  /personJob:
    post:
      tags:
        - person-job-controller
      summary: Создать новую связующую запись сотрудник-должность
      operationId: createPersonJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonJobCreateRequest'
      responses:
        201:
          description: Запись успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonJobResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /personJob/{id}:
    put:
      tags:
        - person-job-controller
      summary: Изменить связующую запись по ID
      operationId: updatePersonJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonJobUpdateRequest'
      responses:
        200:
          description: Запись успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonJobResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /personJob/person/{personId}:
    get:
      tags:
        - person-job-controller
      summary: Найти все должности сотрудника по его ID
      operationId: findPersonJobByPersonId
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID сотрудника
          example: 5
      responses:
        200:
          description: Список должностей сотрудника
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonJobResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /personJob/person/{personId}/current:
    get:
      tags:
        - person-job-controller
      summary: Найти текущую должность сотрудника по его ID
      operationId: findPersonJobByPersonIdAndCurrent
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID сотрудника
          example: 5
      responses:
        200:
          description: Текущая должность сотрудника
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonJobResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
components:
  schemas:
    AcademicDegreesCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название ученой степени
          example: "доктор наук"
        shortName:
          type: string
          description: Сокращенное название
          example: "д.н."
      required:
        - name
...
    AcademicDegreesResponse:
      allOf:
        - $ref: '#/components/schemas/AcademicDegreesCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID учёной степени
          required:
            - id
    AcademicTitlesCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название ученого звания
          example: "профессор"
        shortName:
          type: string
          description: Сокращенное название
          example: "проф."
      required:
        - name
...
    AcademicTitlesResponse:
      allOf:
        - $ref: '#/components/schemas/AcademicTitlesCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID учёного звания
          required:
            - id
    Achievement:
      type: object
      properties:
        title:
          type: string
          description: Название достижения
          example: "Нобелевская премия"
        year:
          type: integer
          format: int32
          description: Год получения достижения
          example: 2020
        description:
          type: string
          description: Описание достижения
          example: "За открытие в области физики"
      required:
        - title
    AchievementResponse:
      allOf:
        - $ref: '#/components/schemas/Achievement'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID достижения
          required:
            - id
...
    EducationSubjectCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название учебного предмета
          example: "экономика"
        shortName:
          type: string
          description: Сокращенное название предмета
          example: "эконом."
      required:
        - name
...
    EducationSubjectResponse:
      allOf:
        - $ref: '#/components/schemas/EducationSubjectCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID научного направления
          required:
            - id
    JobTitleCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название должности
          example: "директор"
      required:
        - name
...
    JobTitleResponse:
      allOf:
        - $ref: '#/components/schemas/JobTitleCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID должности
          required:
            - id
    OutstandingPeopleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID человека
        name:
          type: string
          description: Имя
        surname:
          type: string
          description: Фамилия
        patronymic:
          type: string
          description: Отчество
      required:
        - id
        - name
        - surname
        - patronymic
    OutstandingPersonResponse:
      allOf:
        - $ref: '#/components/schemas/OutstandingPeopleResponse'
        - type: object
          properties:
            gender:
              type: string
              enum: [ NOT_SELECTED, MALE, FEMALE ]
              description: Пол
            yearOfBirth:
              type: integer
              format: int32
              description: Год рождения
            yearOfDeath:
              type: integer
              format: int32
              description: Год смерти
            photoUrl:
              type: string
              description: URL фотографии
            biography:
              type: string
              description: Краткая биография
            academicDegree:
              allOf:
                - $ref: '#/components/schemas/AcademicDegreesResponse'
                - type: object
              nullable: true
              readOnly: true
            academicTitles:
              allOf:
                - $ref: '#/components/schemas/AcademicTitlesResponse'
                - type: object
              nullable: true
              readOnly: true
            educationSubject:
              allOf:
                - $ref: '#/components/schemas/EducationSubjectResponse'
                - type: object
              nullable: true
              readOnly: true
            achievements:
              type: array
              items:
                $ref: '#/components/schemas/AchievementResponse'
              description: Список достижений
              readOnly: true
            jobs:
              type: array
              items:
                $ref: '#/components/schemas/PersonJobResponse'
              description: Список должностей
              readOnly: true
    OutstandingPersonCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Имя
          example: "Геннадий"
        surname:
          type: string
          description: Фамилия
          example: "Хацкевич"
        patronymic:
          type: string
          description: Отчество
          example: "Алексеевич"
        gender:
          type: string
          enum: [ NOT_SELECTED, MALE, FEMALE ]
          default: NOT_SELECTED
          description: Пол
          example: "MALE"
        yearOfBirth:
          type: integer
          format: int32
          description: Год рождения
          example: 1950
        yearOfDeath:
          type: integer
          format: int32
          nullable: true
          description: Год смерти
        photoUrl:
          type: string
          description: URL фотографии
        biography:
          type: string
          description: Краткая биография
        academicTitle:
          type: string
          nullable: true
          description: Название ученого звания
          example: "профессор"
        academicDegree:
          type: string
          nullable: true
          description: Название ученой степени
          example: "доктор"
        educationSubject:
          type: string
          nullable: true
          description: Название учебного предмета
          example: "экономика"
      required:
        - name
        - surname
        - patronymic
    OutstandingPersonUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/OutstandingPersonCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    PersonJobResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID записи
        jobTitle:
          allOf:
            - $ref: '#/components/schemas/JobTitleResponse'
            - type: object
          readOnly: true
        startYear:
          type: integer
          format: int32
          description: Год начала работы
        endYear:
          type: integer
          format: int32
          description: Год окончания работы
        current:
          type: boolean
          description: Является ли должность актуальной
      required:
          - id
    PersonJobCreateRequest:
      type: object
      properties:
        startYear:
          type: integer
          format: int32
          nullable: true
          description: Год начала работы
          example: 2010
        endYear:
          type: integer
          format: int32
          nullable: true
          description: Год окончания работы (не вводить, если должность актуальна)
        current:
          type: boolean
          default: false
          description: Является ли должность актуальной
          example: false
        personId:
          type: integer
          format: int64
          description: Идентификатор человека, к которому относится запись
          example: 1
        jobTitle:
          type: string
          description: Название должности
          example: "ректор"
      required:
        - personId
        - jobTitle
    PersonJobUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/PersonJobCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
        login:
          type: string
          description: Логин пользователя
        role:
          type: string
          description: Роль пользователя
      required:
        - id
        - login
        - role
    UserCreateRequest:
      type: object
      properties:
        login:
          type: string
          maxLength: 50
          description: Логин пользователя
          example: "admin4"
        password:
          type: string
          maxLength: 100
          description: Пароль пользователя
          example: "admin4"
        role:
          type: string
          enum: [ NOT_SELECTED, ADMIN, EDITOR ]
          description: Роль пользователя
          default: NOT_SELECTED
          example: "ADMIN"
      required:
        - login
        - password
    UserUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/UserCreateRequest'
        - type: object
  responses:
    400:
      description: Некорректный запрос
    404:
      description: Ресурс не найден
    500:
      description: Внутренняя ошибка сервера